requirements.txt
aiogram
environs
redis
betterlogging
yookassa
python-dotenv
uvicorn
fastapi
gspread
oauth2client
jinja2
